requirements.txt
Flask==2.2.3
Flask-SocketIO==5.3.2
Flask-Login==0.6.2
Werkzeug==2.2.3
requests==2.28.2
pyserial==3.5
RPi.GPIO==0.7.1
python-dotenv==1.0.0
gunicorn==20.1.0
eventlet==0.33.3
psutil==5.9.4
bluepy==1.3.0
scapy==2.5.0
pexpect==4.8.0 